mysql added

+++ model/user/user.model.go
@@ -2,7 +2,7 @@ package user
 
 // User defines attribute of user
 type User struct {
-	UserID    int    `json:"userId"`
+	UserID    uint64 `json:"userId"`
 	FirstName string `json:"firstName"`
 	LastName  string `json:"lastName"`
 	Email     string `json:"email"`
